refactor: Adjust imports and types
