+++ packages/sdk/scripts/get-logs.ts
@@ -1,5 +1,6 @@
 import {config} from 'dotenv'
 import {Log, PublicClient, createPublicClient, http, parseAbi} from 'viem'
+
 import {HOLOGRAPH_EVENTS} from '../src/constants/events'
 
 config()
