+++ packages/sdk/src/contracts/factory.contract.ts
@@ -1,12 +1,12 @@
 import {Address} from 'abitype'
+import {Hex} from 'viem'
 
 import {Holograph} from '.'
 import {HolographFactoryABI} from '../constants/abi/develop'
-import {BridgeSettings, DeploymentConfig} from '../utils/decoders'
 import {HolographLogger, Config, HolographWallet} from '../services'
 import {HolographByNetworksResponse, getSelectedNetworks} from '../utils/contracts'
-import {GetContractFunctionArgs, HolographBaseContract} from './holograph-base.contract'
-import {Hex} from 'viem'
+import {BridgeSettings, DeploymentConfig, GetContractFunctionArgs} from '../utils/types'
+import {HolographBaseContract} from './holograph-base.contract'
 
 /**
  * @group Contracts
