+++ packages/sdk/src/utils/helpers.ts
@@ -1,7 +1,7 @@
 import {NETWORK_KEY_BY_RPC_URL, NetworkKey, networks} from '@holographxyz/networks'
 
 import {getEnv} from '../config/env.validation'
-import {HolographConfig, NetworkRpc} from '../services'
+import {HolographConfig, NetworkRpc} from './types'
 
 // The function below is for checking where the code is running at (either front-end or server).
 // It is used mainly on the networks setup for the Config service class.
